techpack: camera: cci: Semaphore may sleep in spin lock
When we enable CONFIG_DEBUG_ATOMIC_SLEEP=y, the kernel will do
might_sleep check when we down a semaphore, if the might_sleep
found it is in atomic context, it will eventually call BUG() to
raise kernel panic because sleep is not permitted in atomic
context. What we do now is get a spink lock before down a
semaphore, this is possible to make might_sleep reise a
kernel panic because semaphore might sleep in spin lock. So now
change the spin lock to mutex lock to avoid this situation.

CRs-Fixed: 3269325
Change-Id: I37735dfa615ec7dfe0fc38d04688025badd255d5
Signed-off-by: mingpan <quic_mingpan@quicinc.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
(cherry picked from commit 7f276ba181abf61308f426e1788061e6c00c6cf8)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>